DPL: improve resource manager
* Optimize to use ipc:// on the same node
* Allow specifying resources on the command line. Currently works
  only for a single host.
* Foundations to be able to run the same workflow in multinode
  distributed manner.

+++ Framework/Core/src/WorkflowHelpers.h
@@ -77,7 +77,7 @@ struct DeviceConnectionId {
   size_t consumer;
   size_t timeIndex;
   size_t producerTimeIndex;
-  size_t port;
+  uint16_t port;
 
   bool operator<(const DeviceConnectionId& rhs) const
   {
